ChromeExtension/requirements.txt
websockets==11.0.3
Pillow==10.0.0
